axvisor_api

Module memory

Memory-related API.

This module contains 6 API functions to be implemented:

alloc_frame

alloc_contiguous_frames

dealloc_frame

dealloc_contiguous_frames

phys_to_virt

virt_to_phys

Structs

PhysAddr

A physical memory address.

VirtAddr

A virtual memory address.

Functions

alloc_contiguous_frames

Allocate a number of contiguous frames, with a specified alignment.

alloc_frame

Allocate a frame.

dealloc_contiguous_frames

Deallocate a number of contiguous frames.

dealloc_frame

Deallocate a frame.

phys_to_virt

Convert a physical address to a virtual address.

virt_to_phys

Convert a virtual address to a physical address.

Type Aliases

PhysFrame

A physical frame which will be automatically deallocated when dropped.