axvisor_api::memory

Trait MemoryIf

pub trait MemoryIf {
    // Required methods
    fn alloc_frame() -> Option<PhysAddr>;
    fn alloc_contiguous_frames(
        num_frames: usize,
        frame_align_pow2: usize,
    ) -> Option<PhysAddr>;
    fn dealloc_frame(addr: PhysAddr);
    fn dealloc_contiguous_frames(first_addr: PhysAddr, num_frames: usize);
    fn phys_to_virt(addr: PhysAddr) -> VirtAddr;
    fn virt_to_phys(addr: VirtAddr) -> PhysAddr;
}

The API trait for memory allocation and address translation functionalities.

Required Methods

fn alloc_frame() -> Option<PhysAddr>

Allocate a frame.

fn alloc_contiguous_frames( num_frames: usize, frame_align_pow2: usize, ) -> Option<PhysAddr>

Allocate a number of contiguous frames, with a specified alignment.

fn dealloc_frame(addr: PhysAddr)

Deallocate a frame allocated previously by alloc_frame.

fn dealloc_contiguous_frames(first_addr: PhysAddr, num_frames: usize)

Deallocate a number of contiguous frames allocated previously by alloc_contiguous_frames.

fn phys_to_virt(addr: PhysAddr) -> VirtAddr

Convert a physical address to a virtual address.

fn virt_to_phys(addr: VirtAddr) -> PhysAddr

Convert a virtual address to a physical address.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.